pyproject.toml
[tool.poetry]
name = "app-pfi-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
requests = "^2.32.3"
pytz = "^2024.2"
aiohttp = "^3.10.5"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
